feat: Add schema.Entity() a mixin
